+++ OUDS/Core/Components/Sources/Buttons/OUDSButton.swift
@@ -216,7 +216,7 @@ public struct OUDSButton: View {
     /// else the button text is used.
     private var accessibilityLabel: String {
         if style == .loading {
-            "core_button_loading_a11y".localized()
+            "core_common_loading_a11y".localized()
         } else {
             switch type {
             case let .text(text), let .textAndIcon(text, _), let .icon(_, text):